requirements.txt
numpy==1.15.3
opencv-contrib-python==3.4.3.18
Pillow==3.1.2
